Style edits
@mcleoj2 Awesome job. I could've left the code alone and it would've been totally fine.  It was cool to see some new functions.

Disclaimer here, take what I've put down as suggestions on style and code readability mostly. Some big caveats to the edits: (1) I had no data to test, so may have broken one (or many) things with my edits; (2) We have very different styles, as you will see - one is not 'right' but just different.

A minor but seemingly major change is the structure/folders added to the repo, designed to make the code into a more streamlined version where one file could essentially serve as the runner of the code, and the rest of the source scripts do the work.